requirements.txt
asgiref==3.8.1
Django==4.2.22
sqlparse==0.5.3
typing_extensions==4.14.0
djangorestframework==3.14.0
python-docx==1.1.0
django-cors-headers==4.3.1
Pillow==10.1.0
